requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
xformers<0.0.26
trl
peft
accelerate
bitsandbytes
langchain_community
langchain
streamlit
llama-index
llama-index-embeddings-langchain
llama-index-embeddings-huggingface
llama-index-llms-huggingface
rouge-score
datasets
huggingface_hub[hf_transfer]
sentence-transformers
